// client/src/app/interfaces/form-options.ts
export interface FormOptionsNVs{
  name: string;
  value: undefined | string;
}

export interface FormOptionsNVn{
  name: string;
  value: undefined | number;
}

export interface FormOptionsLVs{
  label: string;
  value: undefined | string;
}

export interface FormOptionsLVn{
  label: string;
  value: undefined | number;
}

export interface FormOptionsNVS {
  name: string;
  value: undefined | number;
  state: string
}

export interface FormOptionsINV {
  id: string;
  name: undefined | string;
  value: undefined | string;
}